azuretrial-master/requirements.txt
fbprophet==0.5
Flask==1.0.2
Cython==0.29.7
pystan==2.19.0.0
matplotlib==3.1.0
numpy==1.16.3
pandas==0.24.2
itsdangerous==0.24
Jinja2>=2.10
MarkupSafe==1.0
Werkzeug==0.15.4 
click==6.7
statsmodels==0.9.0 
more-itertools==7.0.0
scipy==1.2.1
xlrd==1.2.0
XlsxWriter==1.1.7
xlwt==1.3.0
scikit-learn==0.21.2 
openpyxl==2.6.2 
azure-storage-blob==2.0.1 
gunicorn==19.9.0
